Diversas modificações. Criado módulo servicos. Olhar arquivo modificacoes.txt.Criad

+++ src/cfg_tempo_amostra.c
@@ -1,4 +1,4 @@
 
 #include "cfg_tempo_amostra.h"
 
-uint8_t cfg_tempo_amostra_atual;
+uint16_t cfg_tempo_amostra_atual;
